[feat][#67] PlanBoardView에 기존 코드 이전 완료. 이후 hovering 관련 로직 정리 예정
