+++ gridy/Resources/Assets.xcassets/PlanBoardColors/hoveredCell.colorset/Contents.json
@@ -7,7 +7,7 @@
           "alpha" : "1.000",
           "blue" : "0xFF",
           "green" : "0xFF",
-          "red" : "0xFF"
+          "red" : "0xFE"
         }
       },
       "idiom" : "universal"
@@ -25,7 +25,7 @@
           "alpha" : "1.000",
           "blue" : "0xFF",
           "green" : "0xFF",
-          "red" : "0xFF"
+          "red" : "0xFE"
         }
       },
       "idiom" : "universal"
@@ -41,9 +41,9 @@
         "color-space" : "srgb",
         "components" : {
           "alpha" : "1.000",
-          "blue" : "0x49",
-          "green" : "0x49",
-          "red" : "0x49"
+          "blue" : "0x33",
+          "green" : "0x33",
+          "red" : "0x33"
         }
       },
       "idiom" : "universal"
